+++ src/main/scala/edu/ie3/simona/model/participant2/HpModel.scala
@@ -33,16 +33,15 @@ import edu.ie3.simona.model.thermal.ThermalGrid._
 import edu.ie3.simona.ontology.messages.flex.{FlexOptions, MinMaxFlexOptions}
 import edu.ie3.simona.ontology.messages.services.WeatherMessage.WeatherData
 import edu.ie3.simona.service.ServiceType
-import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.quantities.QuantityUtils.RichQuantityDouble
 import edu.ie3.util.scala.quantities.DefaultQuantities.{
   zeroCelsius,
   zeroKW,
   zeroKWh,
 }
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.PowerConversionSimona
 import edu.ie3.util.scala.quantities._
 import squants._
-import squants.energy.Kilowatts
 
 import java.time.ZonedDateTime
 import java.util.UUID
@@ -431,26 +430,18 @@ object HpModel {
       )
     }
 
-    override def create(): HpModel =
+    override def create(): HpModel = {
+      val bmType = input.getType
+
       new HpModel(
         input.getUuid,
         input.getId,
-        Kilovoltamperes(
-          input.getType.getsRated
-            .to(PowerSystemUnits.KILOVOLTAMPERE)
-            .getValue
-            .doubleValue
-        ),
+        bmType.getsRated.toApparent,
         input.getType.getCosPhiRated,
         QControl(input.getqCharacteristics),
-        Kilowatts(
-          input.getType
-            .getpThermal()
-            .to(PowerSystemUnits.KILOWATT)
-            .getValue
-            .doubleValue
-        ),
+        bmType.getpThermal.toSquants,
         ThermalGrid(thermalGrid),
       )
+    }
   }
 }